+++ common/src/main/java/com/st0x0ef/stellaris/client/overlays/OxygenTankOverlay.java
@@ -4,6 +4,9 @@ import com.mojang.blaze3d.systems.RenderSystem;
 import com.st0x0ef.stellaris.Stellaris;
 import com.st0x0ef.stellaris.client.screens.helper.ScreenHelper;
 import com.st0x0ef.stellaris.common.armors.AbstractSpaceArmor;
+import com.st0x0ef.stellaris.common.data_components.OxygenComponent;
+import com.st0x0ef.stellaris.common.registry.DataComponentsRegistry;
+import com.st0x0ef.stellaris.common.utils.OxygenUtils;
 import com.st0x0ef.stellaris.common.utils.Utils;
 import net.minecraft.ChatFormatting;
 import net.minecraft.client.Minecraft;
@@ -28,10 +31,9 @@ public class OxygenTankOverlay {
             ItemStack chest = player.getItemBySlot(EquipmentSlot.CHEST);
             Minecraft mc = Minecraft.getInstance();
 
-            if(chest.getItem() instanceof AbstractSpaceArmor.Chestplate armor) {
-
-                int oxygen = armor.getOxygen(chest);
-                int maxOxygen = armor.getMaxOxygen();
+            if(chest.getItem() instanceof AbstractSpaceArmor.Chestplate) {
+                long oxygen = OxygenUtils.getOxygen(chest);
+                long maxOxygen = OxygenUtils.getOxygenCapacity(chest);
 
                 int x = 5;
                 int y = 5;