+++ changelog.md
@@ -1,3 +1,4 @@
 ## 1.0.7 changelog
 
-- change the recipes of the jet suit (to make it obtainable in survival)
+- change the recipes of the jet suit (now take big oxygen tank)
+- fix jet suit can't take oxygen